refactor: se dió un nombre más adecuado a la dependencia encargada de obtener la información de un usuario autenticado
